Mats changed, ags and tgs are added, coin collect bot sync added

+++ Assets/Scripts/Gadgets/Bots/BotInfo.cs
@@ -7,5 +7,5 @@ public class BotInfo : MonoBehaviour
     public string team;
     public string playerID;
     public string botId;
-    public float botHealth;
+    public int botHealth;
 }
